refactor: Further cleanup code locations and naming
