+++ src/index.ts
@@ -2,7 +2,7 @@ import pact from './pact';
 module.exports = exports = pact;
 export default pact;
 
-export * from './verifier/verifier';
+export * from './verifier';
 
 export * from './server';
