+++ src/pact.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import serverFactory, { Server, ServerOptions } from './server';
 import stubFactory, { Stub, StubOptions } from './stub';
-import verifierFactory from './verifier/verifier';
+import verifierFactory from './verifier';
 import { VerifierOptions } from './verifier/types';
 import messageFactory, { MessageOptions } from './message';
 import publisherFactory, { PublisherOptions } from './publisher';
